Shallow clone the primary checkout and submodules
Inspired by this feature of actions/checkout (which I can't get to work
with an ssh-key in a secret, for some reason) this uses fairly recent
git features to make shallow clones work with submodules.

I removed this command in the case of pull requests:

```
git merge --no-commit 'origin/${{github.base_ref}}'
```

I did't realize that the checkout action was doing merging the target
branch into PRs, are we sure it's a good idea? I we want to keep it I
think I can make it work with a second fetch command.