Use '${{github.sha}}' for merged PR commit
This is extremely thinly documented but it seems that just using the
`sha` provided on a pull_request event will give us the functionality
that was previously implemented here by checking out the pull request
head and merging in the target branch.

This is the best reference I could find:
https://github.com/actions/checkout/issues/15#issuecomment-1302713351

With these changes our code for pull requests is quite similar to that
for branches, we're just configuring the current user name here. Maybe
we should merge those branches and always configure the user?